third_party/blink/web_tests/svg/custom/polyline-invalid-points-expected.txt

CONSOLE ERROR: Error: <polyline> attribute points: Expected number, "…100, 100 0, 100 INF INF".
